Bug fix commit 1: related printing overlap scores
In bug report, 'bug001', 'bug003', 'bug004' is related a function, "hud()"
"hud()" prints scores and keys on bottom of the screen.
But a variable appears bottom of the screen is modificable.
So the variable is replaced a const variable.

The best way is that the function gets const variable parameters about screen size.
But this project's code is not divided.
This bug have to fix the bug in advance.

Therefore, the best way(clean code) is applied more than later.